% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists_to_csv.R
\name{checklists_to_csv}
\alias{checklists_to_csv}
\title{Output *.csv file of the species checklist}
\usage{
checklists_to_csv(checklist, out_dir = "../Output/")
}
\arguments{
\item{checklist}{a \code{\link{list}} created by \code{\link{make_checklists}} with
argument \code{xls = FALSE}}

\item{out_dir}{character string of the file path to the directory in which to save output
comma-separated file(s)}
}
\value{
a comma-separated file for each distinct polygon (i.e., each \code{name} in the
  \code{geo_ebird_df} passed to \code{\link{make_checklists}})
}
\description{
This function generates, for each queried polygon, a *.csv file of the final species
 checklist and abundance classification created by \code{\link{make_checklists}}.  It
 requires that the output of \code{\link{make_checklists}} be output to an R object and
 the argument \code{xls = FALSE}; see example.  This produces a clean file for easy import
 into, e.g., an external database.
}
\examples{
\dontrun{
# Get shapefile
SErefuges <- rgdal::readOGR("../GIS", "refuges", verbose = FALSE, stringsAsFactors = FALSE)

# Query eBird records within Piedmont NWR and 10 km buffer
piedmont <- geo_ebird(SErefuges, which_polys = "Piedmont", buffers = c(0, 10))

# Generate R object containing abundance, occurrence, final checklist, and effort information
# Note to store the output in an R object and specify xls = FALSE
piedmont_cl <- make_checklists(piedmont, xls = FALSE)
checklists_to_csv(piedmont_cl)
}
}

