% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_ebird.R
\name{parse_ebird}
\alias{parse_ebird}
\title{Parse eBird database query and prepare for SQLite database storage}
\usage{
parse_ebird(file)
}
\arguments{
\item{file}{character string of the file path to the tab-delimited eBird .txt file}
}
\value{
a \code{\link{data.frame}}
}
\description{
This function parses the tab-delimited text files (.txt) provided by Cornell.  Only fields
relevant to the calculation of species-level geographic occurrence are preserved, and
filters are applied to remove eBird records that were not identified to at least the species
level, duplicate records present in group checklists and non-approved records.  Records for
species with an unspecified count of individuals are assigne a minimum value (i.e., 1) to
indicate occurrence.
}

