% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ebird_phen.R
\name{plot_ebird_phen}
\alias{plot_ebird_phen}
\title{Create species phenology plots (monthly) of eBird species abundance and occurrence (# checklists).}
\usage{
plot_ebird_phen(geo_ebird_df, species = NULL, complete_only = TRUE)
}
\arguments{
\item{geo_ebird_df}{a \code{\link{data.frame}} created by \code{\link{geo_ebird}}}

\item{species}{a character string of length = 1 indicating the species to visualize.  If NULL
(the default), the user is presented with a list of available species from which to choose.
Use the accepted American Ornithologists' Union name \link{http://checklist.aou.org/taxa/},
insensitive to capitalization.}

\item{complete_only}{logical indicating whether only eBird checklists designating as reporting
all species observed should be used (default = TRUE)}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
This function generates, for each queried polygon, a phenology plot of eBird sightings
 (proportion of checklists with species observed) by month. The number of checklists
 present for each month is also indicated graphically.
}

