% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/store_ebird.R
\name{store_ebird}
\alias{store_ebird}
\title{Send filtered eBird data frame to SQLite database.}
\usage{
store_ebird(df, db = "../Data/SE_eBird.sqlite",
  table_name = "SE_eBird_Aug2015")
}
\arguments{
\item{df}{a \code{\link{data.frame}} created by \code{parse_ebird}}

\item{db}{character string containing the file path to the appropriate SQLite database}

\item{table_name}{character string specifying a DBMS table name}
}
\description{
This function appends a \code{\link{data.frame}} created by \code{parse_ebird}
to an existing SQLite database.
}

