% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ebird_phen.R
\name{plot_ebird_phen}
\alias{plot_ebird_phen}
\title{Create species phenology plots (monthly) of eBird species abundance and occurrence (# checklists).}
\usage{
plot_ebird_phen(geo_ebird_df, species = NULL, which_polys = NULL,
  complete_only = TRUE)
}
\arguments{
\item{geo_ebird_df}{a \code{\link{data.frame}} created by \code{\link{geo_ebird}}}

\item{species}{a character string of length = 1 indicating the species to visualize.  If NULL
(the default), the user is presented with a list of available species from which to choose.
Use the accepted American Ornithologists' Union name (see http://checklist.aou.org/taxa),
insensitive to capitalization (e.g., "PuRpLe MaRtIn" works).}

\item{which_polys}{character vector of the names of polygons to plot.  Default (NULL) is to
generate a plot showing all polygons.  This argument is insensitive to capitalization.}

\item{complete_only}{logical indicating whether only eBird checklists designated as reporting
all species observed (i.e., complete checklists; default = TRUE) should be used to calculate
proportion of checklists or whether all available checklists should be used (FALSE)}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
This function generates, for each queried polygon, a phenology plot of eBird sightings by month.
 The user can choose between displaying the proportion of only complete checklists or all
 available checklists. The number of checklists is indicated graphically.
}
\examples{
\dontrun{
# Get shapefile
SErefuges <- rgdal::readOGR("../GIS", "refuges", verbose = FALSE, stringsAsFactors = FALSE)

# Query eBird records within Piedmont NWR and 10 km buffer
piedmont <- geo_ebird(SErefuges, which_polys = "Piedmont", buffers = c(0, 10))

# Plot phenology of Pileated Woodpecker
# Species not case sensitive
plot_ebird_phen(piedmont, species = "PiLEAted WOOdPECKer")
}
}

