% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_ebird.R
\name{store_ebird}
\alias{store_ebird}
\title{Send filtered eBird data frame to SQLite database.}
\usage{
store_ebird(df, file = NULL, db = "../Data/SE_eBird.sqlite",
  table_name = "SE_eBird_Aug2015")
}
\arguments{
\item{df}{a \code{\link{data.frame}} created by \code{parse_ebird}}

\item{file}{character string of the file path to the tab-delimited eBird .txt file; if
specified, \code{df} is ignored. Useful for combining the parsing and storing of eBird
records in a single step}

\item{db}{character string containing the file path to the appropriate SQLite database}

\item{table_name}{character string specifying a DBMS table name}
}
\description{
This function appends a \code{\link{data.frame}} created by \code{\link{parse_ebird}}
to an existing SQLite database.
}
\details{
This requires an existing SQLite database that is ready to receive the output of
 \code{\link{parse_ebird}}.  See example below for how the local database was set up
 for southeast refuges.
}
\examples{
\dontrun{
# Set up eBird SQLite database
library(sqldf)
db <- dbConnect(SQLite(), dbname="./Data/SE_eBird.sqlite")

dbSendQuery(conn = db,
            "CREATE TABLE SE_eBird_Aug2015
            (id TEXT,
            tax_order REAL,
            category TEXT,
            common_name TEXT,
            sci_name TEXT,
            count INTEGER,
            county TEXT,
            lat REAL,
            lon REAL,
            date INTEGER,
            checklist TEXT,
            all_spp INTEGER)")

ga_birds <- parse_ebird("./Data/ebd_US-GA_relAug-2015.txt")
store_ebird(ga_birds)

# Close connection
dbDisconnect(db)
}
}

